chore: ONLY place changing of the functions for easier debug flow, tester on (145/146) because of export outfile fd handling
