+++ parser/clean_unnecessary_fds.c
@@ -6,45 +6,15 @@
 /*   By: zslowian <zslowian@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/27 20:21:20 by zslowian          #+#    #+#             */
-/*   Updated: 2025/05/03 16:44:34 by zslowian         ###   ########.fr       */
+/*   Updated: 2025/05/04 20:45:11 by zslowian         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-bool		process_env_variable(char *env_var, t_list **list);
 void		add_io_heredoc_data(t_global *g, t_command *cmd, char *delimiter);
 static void	ft_check_preceedent_heredocs(t_list *cmd_io);
 char		*get_heredoc_name(void);
-
-bool	process_env_variable(char *env_var, t_list **list)
-{
-	t_minishell_env	*content;
-
-	content = ft_calloc(sizeof(t_minishell_env), 1);
-	if (!content)
-	{
-		ft_lstclear(list, ft_clear_minishell_env);
-		return (false);
-	}
-	content->name_value = ft_calloc(sizeof(char *), 2);
-	if (!content->name_value)
-	{
-		ft_lstclear(list, ft_clear_minishell_env);
-		return (false);
-	}
-	ft_split_env_variable(env_var, &content->name_value[0],
-		&content->name_value[1]);
-	if (!content->name_value[0] || !content->name_value[1])
-	{
-		free(content);
-		ft_lstclear(list, ft_clear_minishell_env);
-		return (false);
-	}
-	if (ft_strcmp(content->name_value[0], "_"))
-		content->export = true;
-	return (ft_lstadd_back(list, ft_lstnew(content)), true);
-}
 
 void	add_io_heredoc_data(t_global *g, t_command *cmd, char *delimiter)
 {
